% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{cat2}
\alias{cat2}
\title{Resolve compatibility issues when using cat2}
\usage{
cat2(..., end = "\\n", level = "DEBUG", print_level = FALSE,
  pal = list(DEBUG = "grey60", INFO = "#1d9f34", WARNING = "#ec942c",
  ERROR = "#f02c2c", FATAL = "#763053", DEFAULT = "#000000"))
}
\arguments{
\item{...}{objects to be printed}

\item{end}{end of print, default is with a new line}

\item{level}{six levels, check pal}

\item{print_level}{logical, prepend level information to the output}

\item{pal}{a list of palette indicating the color of each level}
}
\description{
Resolve compatibility issues when using cat2
}
