% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_index.R
\name{cluster-index}
\alias{cluster-index}
\alias{cluster_index_methods}
\alias{cluster_index}
\alias{best_criterion}
\alias{best_nclusters}
\title{Calculate clustering index}
\usage{
cluster_index_methods(simplify = TRUE)

cluster_index(
  x,
  cluster,
  dist,
  dist_args = list(),
  methods = "default",
  check = TRUE
)

best_criterion(index)

best_nclusters(index)
}
\arguments{
\item{simplify}{whether to return unlisted method names}

\item{x}{numerical matrix of original data (dimension: sample size times
number of variate)}

\item{cluster}{integer vector or matrix of the clustering results. 
see 'Details'}

\item{dist}{distance matrix of \code{x}, used by \code{'silhouette'} and 
\code{'c_index'}. If missing, then the value will be automatically 
calculated from \code{\link[stats]{dist}} function with additional arguments
specified by \code{dist_args}}

\item{dist_args}{see \code{dist}}

\item{methods}{cluster index criterion; a full list can be obtained via
\code{cluster_index_methods()}. Alternatively, you can specify 
\code{methods='default'} or \code{methods='all'} to apply default or all
criterion accordingly.}

\item{check}{see 'Details'}

\item{index}{the table returned by function \code{cluster_index}}
}
\description{
Calculate clustering index
}
\details{
When \code{cluster} is an integer vector, its length must be equal
to the sample size. The observations with the same integer are considered 
to be in the same cluster. If \code{cluster} is a matrix, then its row must 
be consistent with the sample size.

For \code{'ball_hall'}, \code{'trace_w'}, \code{'trace_wib'}, and 
\code{'log_ss_ratio'}, \code{cluster} must be a matrix so 
\code{best_criterion} and \code{best_nclusters} can work properly. 
In addition, the number of clusters indicated by \code{cluster} columns
must be increasing. The \code{cluster_index} will check \code{cluster} by 
default, which can be disabled by specifying \code{check=FALSE}
}
