% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_mse.R
\name{cluster_mse}
\alias{cluster_mse}
\title{Calculate cluster centroid and mean standard error}
\usage{
cluster_mse(indata_plot, cluster)
}
\description{
Calculate cluster centroid and mean standard error
}
