% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cluster_table.R
\name{convert_cluster_table}
\alias{convert_cluster_table}
\title{Convert 'RAVE' cluster results into a readable table}
\usage{
convert_cluster_table(x, split_by = "A", var = "C", value = "B")
}
\arguments{
\item{x}{a list of cluster results}

\item{split_by}{variable name to be split by; for example, \code{'Subject'}}

\item{var}{variable name indicating the cluster results; for example, 
\code{'Cluster'}}

\item{value}{the name of variable to concatenate into a single string or cell
value; for example, \code{'Electrode'}}
}
\value{
A \code{\link{data.frame}} that is readable
}
\description{
Convert 'RAVE' cluster results into a readable table
}
