% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{rave-cluster-data-import}
\alias{rave-cluster-data-import}
\alias{scan_power_explorer_exports}
\alias{find_source}
\alias{import_configurations}
\title{Data import functions used to locate data exported by power explorer
module.}
\usage{
scan_power_explorer_exports(search_paths, pattern = "\\\\.fst$")

find_source(search_paths, fname)

import_configurations(search_paths, file_names)
}
\arguments{
\item{search_paths}{directories to search for data}

\item{pattern}{search regular expression pattern}

\item{fname}{character vector of length 1: file name to locate within \code{search_paths}}

\item{file_names}{vector of \code{fname} to import}
}
\value{
potential file names to be loaded
}
\description{
Data import functions used to locate data exported by power explorer
module.
}
