% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{raveclusters-globals}
\alias{raveclusters-globals}
\alias{RAVE_ROI_KEY}
\alias{shiny_cex.main}
\title{Global variables used by 'RAVE' clustering module}
\format{
An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
RAVE_ROI_KEY

shiny_cex.main
}
\description{
Global variables used by 'RAVE' clustering module
}
\keyword{datasets}
